requirements.txt
boto3==1.33.7
botocore==1.33.7
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.8.2
schedule==1.2.1
six==1.16.0
urllib3==1.26.18
